+++ crates/lsp/src/import_resolution.rs
@@ -108,11 +108,6 @@ impl<'a> ImportCursor<'a> {
         None
     }
 
-    #[cfg(test)]
-    pub(crate) fn path_prefix(&self) -> &'a str {
-        self.path_prefix
-    }
-
     pub(crate) fn decoded_path_prefix(&self) -> Option<String> {
         let mut invalid_escape = false;
         let bytes = try_parse_string_literal(self.path_prefix, StrKind::Str, |_, _| {
@@ -512,578 +507,5 @@ fn insert_candidate(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{ImportCursor, ImportResolutionContext, ImportResolver, MAX_IMPORT_CANDIDATES};
-    use crate::{test_support::TestProject, workspace::Workspace};
-
-    fn marked_source(source: &str) -> (String, usize) {
-        let cursor = source.find("$0").unwrap();
-        (source.replacen("$0", "", 1), cursor)
-    }
-
-    #[test]
-    fn import_cursor_recovers_an_unterminated_path_without_consuming_the_next_line() {
-        let (source, offset) = marked_source("import \"./Dep$0\ncontract C {}\n");
-
-        let cursor = ImportCursor::at(&source, offset).unwrap();
-
-        assert_eq!(cursor.path_prefix(), "./Dep");
-        assert_eq!(&source[cursor.replacement_range()], "./Dep");
-        assert_eq!(cursor.complete_path(), None);
-    }
-
-    #[test]
-    fn import_cursor_does_not_use_a_quote_on_the_next_line_as_the_terminator() {
-        let (source, offset) =
-            marked_source("import \"./Dep$0\ncontract C { string value = \"ordinary\"; }\n");
-
-        let cursor = ImportCursor::at(&source, offset).unwrap();
-
-        assert_eq!(cursor.path_prefix(), "./Dep");
-        assert_eq!(&source[cursor.replacement_range()], "./Dep");
-        assert_eq!(cursor.complete_path(), None);
-    }
-
-    #[test]
-    fn import_cursor_treats_backslash_bare_cr_as_an_unescaped_line_break() {
-        let (source, offset) =
-            marked_source("import \"./Dep$0\\\rcontract C { string value = \"ordinary\"; }\r");
-
-        let cursor = ImportCursor::at(&source, offset).unwrap();
-
-        assert_eq!(cursor.path_prefix(), "./Dep");
-        assert_eq!(&source[cursor.replacement_range()], "./Dep");
-        assert_eq!(cursor.complete_path(), None);
-    }
-
-    #[test]
-    fn import_cursor_does_not_replace_unknown_unterminated_suffixes() {
-        let (source, offset) = marked_source("import \"./Dep$0 contract C {}");
-
-        let cursor = ImportCursor::at(&source, offset).unwrap();
-
-        assert_eq!(cursor.path_prefix(), "./Dep");
-        assert_eq!(&source[cursor.replacement_range()], "./Dep");
-        assert_eq!(cursor.complete_path(), None);
-    }
-
-    #[test]
-    fn import_cursor_supports_escaped_line_continuations() {
-        for newline in ["\n", "\r\n"] {
-            let marked = format!("import \"./nested/\\{newline}    Tar$0get.sol\";");
-            let (source, offset) = marked_source(&marked);
-
-            let cursor = ImportCursor::at(&source, offset).unwrap();
-
-            assert_eq!(cursor.decoded_path_prefix().as_deref(), Some("./nested/Tar"));
-            assert_eq!(
-                &source[cursor.replacement_range()],
-                format!("./nested/\\{newline}    Target.sol")
-            );
-            assert_eq!(
-                cursor.complete_path(),
-                Some(format!("./nested/\\{newline}    Target.sol").as_str())
-            );
-        }
-    }
-
-    #[test]
-    fn import_cursor_recovers_after_an_import_missing_its_semicolon() {
-        let (source, offset) = marked_source("import \"./First.sol\"\nimport \"./Second$0.sol\";");
-
-        let cursor = ImportCursor::at(&source, offset).unwrap();
-
-        assert_eq!(cursor.path_prefix(), "./Second");
-        assert_eq!(&source[cursor.replacement_range()], "./Second.sol");
-        assert_eq!(cursor.complete_path(), Some("./Second.sol"));
-    }
-
-    #[test]
-    fn import_cursor_recognizes_import_forms_and_rejects_ordinary_strings() {
-        for marked in [
-            "import \"./Dep$0\";",
-            "import \"./Dep$0\" as Dependency;",
-            "import * as Dependency from \"./Dep$0\";",
-            "import {A, B as C} from \"./Dep$0\";",
-            "import {\n    A\n} from './Dep$0';",
-        ] {
-            let (source, offset) = marked_source(marked);
-            let cursor = ImportCursor::at(&source, offset)
-                .unwrap_or_else(|| panic!("expected an import cursor for {source:?}"));
-
-            assert_eq!(cursor.path_prefix(), "./Dep");
-            assert_eq!(&source[cursor.replacement_range()], "./Dep");
-            assert_eq!(cursor.complete_path(), Some("./Dep"));
-        }
-
-        for marked in [
-            "string constant VALUE = \"./Dep$0\";",
-            "contract C { function f() external { string memory x = \"./Dep$0\"; } }",
-        ] {
-            let (source, offset) = marked_source(marked);
-            assert_eq!(ImportCursor::at(&source, offset), None, "source: {source:?}");
-        }
-    }
-
-    #[test]
-    fn resolver_completes_only_one_relative_directory_level_and_includes_overlay_paths() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-
-            //- /src/Main.sol
-            import "./";
-
-            //- /src/Local.sol
-            contract Local {}
-
-            //- /src/nested/OnDisk.sol
-            contract OnDisk {}
-
-            //- /src/README.md
-            not Solidity
-            "#,
-        );
-        let config = project.config();
-        let importer = project.path("/src/Main.sol");
-        let context = config.import_resolution_context(&importer).unwrap();
-        let overlay =
-            [project.path("/src/Unsaved.sol"), project.path("/src/virtual/OnlyInOverlay.sol")];
-
-        let completion = ImportResolver::new(context, &overlay).complete(&importer, "./");
-        let candidates = completion
-            .candidates()
-            .iter()
-            .map(|candidate| candidate.import_path())
-            .collect::<Vec<_>>();
-
-        assert_eq!(
-            candidates,
-            vec!["./Local.sol", "./Main.sol", "./Unsaved.sol", "./nested/", "./virtual/"]
-        );
-        assert!(!completion.is_incomplete());
-    }
-
-    #[test]
-    fn resolver_completes_bare_relative_directory_segments() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-
-            //- /Root.sol
-            contract Root {}
-
-            //- /src/Main.sol
-            import ".";
-
-            //- /src/Local.sol
-            contract Local {}
-            "#,
-        );
-        let config = project.config();
-        let importer = project.path("/src/Main.sol");
-        let context = config.import_resolution_context(&importer).unwrap();
-        let resolver = ImportResolver::new(context, &[]);
-
-        for (prefix, expected) in [(".", "./"), ("..", "../"), ("./.", "././"), ("../..", "../../")]
-        {
-            let completion = resolver.complete(&importer, prefix);
-            assert_eq!(
-                completion
-                    .candidates()
-                    .iter()
-                    .map(|candidate| candidate.import_path())
-                    .collect::<Vec<_>>(),
-                [expected],
-                "prefix: {prefix:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn resolver_keeps_dotfiles_beside_relative_directory_continuations() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-
-            //- /src/Main.sol
-            import ".";
-
-            //- /src/.b.sol
-            contract SingleDot {}
-
-            //- /src/..b.sol
-            contract DoubleDot {}
-            "#,
-        );
-        let config = project.config();
-        let importer = project.path("/src/Main.sol");
-        let context = config.import_resolution_context(&importer).unwrap();
-        let resolver = ImportResolver::new(context, &[]);
-
-        for (prefix, expected) in
-            [(".", vec!["..b.sol", "./", ".b.sol"]), ("..", vec!["../", "..b.sol"])]
-        {
-            let completion = resolver.complete(&importer, prefix);
-            assert_eq!(
-                completion
-                    .candidates()
-                    .iter()
-                    .map(|candidate| candidate.import_path())
-                    .collect::<Vec<_>>(),
-                expected,
-                "prefix: {prefix:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn resolver_resolves_an_overlay_only_import() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-
-            //- /src/Main.sol
-            import "./Unsaved.sol";
-            "#,
-        );
-        let config = project.config();
-        let importer = project.path("/src/Main.sol");
-        let target = project.path("/src/Unsaved.sol");
-        let context = config.import_resolution_context(&importer).unwrap();
-        let overlay = [target.clone()];
-
-        let resolved = ImportResolver::new(context, &overlay).resolve(&importer, "./Unsaved.sol");
-
-        assert_eq!(resolved, Some(target));
-    }
-
-    #[test]
-    fn resolver_rejects_ambiguous_exact_imports() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            libs = ["lib", "vendor"]
-
-            //- /src/Main.sol
-            import "pkg/Target.sol";
-
-            //- /lib/pkg/Target.sol
-            contract LibTarget {}
-
-            //- /vendor/pkg/Target.sol
-            contract VendorTarget {}
-            "#,
-        );
-        let config = project.config();
-        let importer = project.path("/src/Main.sol");
-        let context = config.import_resolution_context(&importer).unwrap();
-
-        let resolved = ImportResolver::new(context, &[]).resolve(&importer, "pkg/Target.sol");
-
-        assert_eq!(resolved, None);
-    }
-
-    #[test]
-    fn resolver_normalizes_the_foundry_root_before_applying_context_remappings() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /container/.keep
-
-            //- /project/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["src:pkg/=lib/"]
-
-            //- /project/src/Main.sol
-            import "pkg/Target.sol";
-
-            //- /project/lib/Target.sol
-            contract Target {}
-            "#,
-        );
-        let manifest = project.path("/container/../project/foundry.toml");
-        let workspaces = [Workspace::load_foundry(manifest).unwrap()];
-        let importer = project.path("/project/src/Main.sol");
-        let context = ImportResolutionContext::for_workspaces(&workspaces, &importer).unwrap();
-
-        let resolved = ImportResolver::new(context, &[]).resolve(&importer, "pkg/Target.sol");
-
-        assert_eq!(resolved, Some(project.path("/project/lib/Target.sol")));
-    }
-
-    #[test]
-    fn resolver_caps_candidates_and_marks_the_result_incomplete() {
-        let project = TestProject::new();
-        project.write_file("/foundry.toml", "");
-        project.write_file("/src/Main.sol", "import \"./\";");
-        for index in 0..MAX_IMPORT_CANDIDATES {
-            project.write_file(&format!("/src/Candidate{index:03}.sol"), "");
-        }
-        let config = project.config();
-        let importer = project.path("/src/Main.sol");
-        let context = config.import_resolution_context(&importer).unwrap();
-
-        let completion = ImportResolver::new(context, &[]).complete(&importer, "./");
-
-        assert_eq!(completion.candidates().len(), MAX_IMPORT_CANDIDATES);
-        assert_eq!(completion.candidates()[0].import_path(), "./Candidate000.sol");
-        assert_eq!(
-            completion.candidates()[MAX_IMPORT_CANDIDATES - 1].import_path(),
-            "./Candidate255.sol"
-        );
-        assert!(completion.is_incomplete());
-    }
-
-    #[test]
-    fn config_selects_the_deepest_import_resolution_context() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/outer/src/"]
-
-            //- /src/Outer.sol
-            contract Outer {}
-
-            //- /packages/app/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/inner/src/"]
-
-            //- /packages/app/src/Inner.sol
-            contract Inner {}
-            "#,
-        );
-        let config = project.config();
-
-        let outer = config.import_resolution_context(&project.path("/src/Outer.sol")).unwrap();
-        let inner = config
-            .import_resolution_context(&project.path("/packages/app/lib/pkg/Overlay.sol"))
-            .unwrap();
-
-        assert_eq!(outer.workspace_root(), project.root());
-        assert_eq!(inner.workspace_root(), project.path("/packages/app"));
-        assert_eq!(
-            outer
-                .compile_opts()
-                .import_remappings
-                .iter()
-                .find(|remapping| remapping.prefix == "pkg/")
-                .unwrap()
-                .path,
-            "lib/outer/src/"
-        );
-        assert_eq!(
-            inner
-                .compile_opts()
-                .import_remappings
-                .iter()
-                .find(|remapping| remapping.prefix == "pkg/")
-                .unwrap()
-                .path,
-            "lib/inner/src/"
-        );
-    }
-
-    #[test]
-    fn config_isolates_import_contexts_across_workspace_roots() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /first/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/first/"]
-
-            //- /first/src/Main.sol
-            contract First {}
-
-            //- /second/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/second/"]
-
-            //- /second/src/Main.sol
-            contract Second {}
-            "#,
-        );
-        let config = project.config_with_roots(&["/first", "/second"]);
-
-        let first = config.import_resolution_context(&project.path("/first/src/Main.sol")).unwrap();
-        let second =
-            config.import_resolution_context(&project.path("/second/src/Main.sol")).unwrap();
-
-        assert_eq!(first.workspace_root(), project.path("/first"));
-        assert_eq!(second.workspace_root(), project.path("/second"));
-        assert_eq!(first.compile_opts().import_remappings[0].path, "lib/first/");
-        assert_eq!(second.compile_opts().import_remappings[0].path, "lib/second/");
-    }
-
-    #[test]
-    fn config_owns_external_source_roots_without_falling_back_to_the_first_workspace() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /first/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/first/"]
-
-            //- /first/src/Main.sol
-            contract First {}
-
-            //- /second/foundry.toml
-            [profile.default]
-            src = "../shared"
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/second/"]
-
-            //- /shared/Main.sol
-            contract Shared {}
-            "#,
-        );
-        let config = project.config_with_roots(&["/first", "/second"]);
-
-        let context = config.import_resolution_context(&project.path("/shared/Main.sol")).unwrap();
-
-        assert_eq!(context.workspace_root(), project.path("/second"));
-        assert_eq!(context.compile_opts().import_remappings[0].path, "lib/second/");
-    }
-
-    #[test]
-    fn config_prefers_a_deeper_external_source_root_over_an_ancestor_base_path() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /outer/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/outer/"]
-
-            //- /outer/packages/app/foundry.toml
-            [profile.default]
-            src = "../../shared"
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/inner/"]
-
-            //- /outer/shared/Main.sol
-            contract Shared {}
-            "#,
-        );
-        let config = project.config_with_roots(&["/outer"]);
-
-        let context =
-            config.import_resolution_context(&project.path("/outer/shared/Main.sol")).unwrap();
-
-        assert_eq!(context.workspace_root(), project.path("/outer/packages/app"));
-        assert_eq!(context.compile_opts().import_remappings[0].path, "lib/inner/");
-    }
-
-    #[test]
-    fn config_owns_external_import_only_roots_without_cross_project_contamination() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /first/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/first/"]
-
-            //- /first/src/Main.sol
-            contract First {}
-
-            //- /second/foundry.toml
-            [profile.default]
-            libs = ["../dependencies"]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/second/"]
-
-            //- /second/src/Main.sol
-            contract Second {}
-
-            //- /dependencies/pkg/Dependency.sol
-            contract Dependency {}
-            "#,
-        );
-        let config = project.config_with_roots(&["/first", "/second"]);
-
-        let context = config
-            .import_resolution_context(&project.path("/dependencies/pkg/Overlay.sol"))
-            .unwrap();
-
-        assert_eq!(context.workspace_root(), project.path("/second"));
-        assert_eq!(context.compile_opts().import_remappings[0].path, "lib/second/");
-    }
-
-    #[test]
-    fn config_owns_out_of_base_remapping_targets() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /project/foundry.toml
-            [profile.default]
-            auto_detect_remappings = false
-            remappings = ["pkg/=../shared/"]
-
-            //- /project/src/Main.sol
-            import "pkg/Dependency.sol";
-
-            //- /shared/Dependency.sol
-            contract Dependency {}
-            "#,
-        );
-        let config = project.config_with_roots(&["/project"]);
-
-        let context =
-            config.import_resolution_context(&project.path("/shared/Dependency.sol")).unwrap();
-
-        assert_eq!(context.workspace_root(), project.path("/project"));
-        assert_eq!(context.compile_opts().import_remappings[0].path, "../shared/");
-    }
-
-    #[test]
-    fn config_does_not_guess_an_import_context_for_an_unowned_path() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /first/foundry.toml
-
-            //- /first/src/Main.sol
-            contract First {}
-
-            //- /second/foundry.toml
-
-            //- /second/src/Main.sol
-            contract Second {}
-            "#,
-        );
-        let config = project.config_with_roots(&["/first", "/second"]);
-
-        assert!(config.import_resolution_context(&project.path("/unowned/Overlay.sol")).is_none());
-    }
-
-    #[test]
-    fn config_rejects_ambiguous_shared_import_contexts() {
-        let project = TestProject::from_fixture(
-            r#"
-            //- /first/foundry.toml
-            [profile.default]
-            libs = ["../shared"]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/first/"]
-
-            //- /second/foundry.toml
-            [profile.default]
-            libs = ["../shared"]
-            auto_detect_remappings = false
-            remappings = ["pkg/=lib/second/"]
-
-            //- /shared/Dependency.sol
-            contract Dependency {}
-            "#,
-        );
-        let config = project.config_with_roots(&["/first", "/second"]);
-
-        assert!(
-            config.import_resolution_context(&project.path("/shared/Dependency.sol")).is_none()
-        );
-    }
-}
+#[path = "import_resolution/tests/mod.rs"]
+mod tests;
