+++ crates/lsp/src/import_resolution/tests/mod.rs
@@ -1,0 +1,3 @@
+mod context;
+mod cursor;
+mod resolver;
